SLP: Add option to disable binding to well-known port
In commit 4985722c, changes were made to cause the SLP socket to be
bound to an ephemeral local port to enable multiple SLP sockets to
co-exist.

As part of this change, all SLP sockets gained functionality causing
them to *also* bind to the SLP well-known port for the purposes of
receiving multicasts, which previously would not have happened for
sockets created by agent implementations of type `SlpUserAgent`.

This causes issues in environments (such as Linux) where binding to
well-known port numbers is a privileged operation, meaning that starting
the `SlpDeviceManager` on these platforms as a non-privileged user would
cause `SocketExceptions` to be thrown.

This commit adds a property to the `SlpDeviceManager` which allows
callers to disable opening of the well-known port.