SLP: Ensure multicast group is always joined
For SLP to function, it is necessary for the agent to join the UDP
multicast group to receive (amongst others) service request messages.
However, only the socket bound to the well-known port was joining the
multicast group, and if this was disabled then there would be no socket
bound to receive traffic from the multicast group, preventing discovery
from functioning.

This commit causes the ephemeral bound port to join the multicast group
if the well-known port is not being opened, allowing SLP discovery to
function in this case.

Related TFS: #87348